# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Talib MR - University Student Platform for Mauritania
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses for current student
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: semesterId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
                required: [success, data]
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course detail
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/CourseDetail"
                required: [success, data]

  # ─── FILES ───────────────────────────────────────────────────────────────────
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List academic files
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [lecture, td, tp, summary, exam, correction, book, other]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Files list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AcademicFile"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required: [success, data]
  /files/{fileId}/favorite:
    post:
      operationId: toggleFileFavorite
      tags: [files]
      summary: Toggle favorite status of a file
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      isFavorited:
                        type: boolean
                    required: [isFavorited]
                required: [success, data]

  # ─── ANNOUNCEMENTS ───────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List upcoming events
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                required: [success, data]
  /events/{eventId}/register:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register for an event
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CLUBS ───────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
                required: [success, data, unreadCount]
  /notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── AI ─────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ─── AUTH SCHEMAS ────────────────────────────────────────────────────────
...
    Course:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        semester:
          type: string
        professorId:
          type: string
          nullable: true
        professorName:
          type: string
          nullable: true
        fileCount:
          type: integer
        assignmentCount:
          type: integer
      required: [id, name, semester, fileCount, assignmentCount]
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            recentFiles:
              type: array
              items:
                $ref: "#/components/schemas/AcademicFile"
            recentAnnouncements:
              type: array
              items:
                $ref: "#/components/schemas/Announcement"
            upcomingAssignments:
              type: array
              items:
                $ref: "#/components/schemas/Assignment"
            upcomingExams:
              type: array
              items:
                $ref: "#/components/schemas/Exam"
          required: [recentFiles, recentAnnouncements, upcomingAssignments, upcomingExams]

    # ─── FILE SCHEMAS ─────────────────────────────────────────────────────────
    AcademicFile:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        fileType:
          type: string
          enum: [lecture, td, tp, summary, exam, correction, book, other]
        mimeType:
          type: string
        fileSize:
          type: integer
        fileUrl:
          type: string
        courseId:
          type: string
          nullable: true
        courseName:
          type: string
          nullable: true
        uploadedById:
          type: string
        uploaderName:
          type: string
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        downloadCount:
          type: integer
        isFavorited:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, fileType, mimeType, fileSize, fileUrl, uploadedById, uploaderName, approvalStatus, downloadCount, isFavorited, createdAt]

    # ─── ANNOUNCEMENT SCHEMAS ─────────────────────────────────────────────────
    Announcement:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [normal, important, urgent]
        scope:
          type: string
          enum: [global, university, faculty, department, level, group, course]
        createdById:
          type: string
        createdByName:
          type: string
        courseName:
          type: string
          nullable: true
        isRead:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, content, priority, scope, createdById, createdByName, isRead, createdAt]

    # ─── TIMETABLE SCHEMAS ────────────────────────────────────────────────────
...
    Assignment:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        courseId:
          type: string
        courseName:
          type: string
        deadline:
          type: string
          format: date-time
        submissionStatus:
          type: string
          enum: [not_submitted, submitted, late, reviewed]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, courseId, courseName, deadline, createdAt]
...
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        courseId:
          type: string
        courseName:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
          nullable: true
        room:
          type: string
          nullable: true
        type:
          type: string
          enum: [midterm, final, makeup, test, other]
        createdAt:
          type: string
          format: date-time
      required: [id, title, courseId, courseName, date, type, createdAt]

    # ─── COMMUNITY SCHEMAS ────────────────────────────────────────────────────
...
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [university, club, training, competition, workshop, conference, other]
        location:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        isRegistered:
          type: boolean
        registrationCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, type, startDate, isRegistered, registrationCount, createdAt]

    # ─── CLUB SCHEMAS ─────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, type, title, body, isRead, createdAt]

    # ─── DASHBOARD SCHEMAS ────────────────────────────────────────────────────
...